Break sales analyst into modules.
